Implement Engine as top-level scaffolding
There are no tests for this (especially no "test first" tests),
since it is close to the infrastructure and platform and only
holds everything together relatively loosely.

Some architectural decisions, like not abstracting the renderer
and not injecting dependencies, are intentional in order to be able
to create additional kata exercises, like discussing possible
architectural changes or problems.