Parse player input and delegate it to the rule engine
The game can now already be played, as long as both players know
the rules and abide by them. The rules are not yet enforced by the
game. All chess pieces can now be moved arbitrarily (like on a real
chess board). Invalid move commands lead to errors which are not yet
being handled.